fix: continue using abstract class for createJwtPayload in order to preserve attribute check for cp
